raspi4b: Modify DTB to set DMA ranges
There is a hardware bug in the Raspberry Pi 4B that is documented
in the pcie-brcmstb driver in the Linux kernel.

The "firmware has to dynamically edit dma-ranges due to a bug on the
PCIe controller inegration".  Since there is no firmware running
inside QEMU, we must instead modify any provided DTBs to accomplish
the same effect.

Signed-off-by: Rob Woolley <rob.woolley@windriver.com>